[2063] Move upstream_connector out from follower directory Author: Miklos Szegedi
